develop(1219-secretsmanager-rotation): add lambda result page
* fix(1219-dynamodb-resolve-host-exception): fix resolve exception, fix…

* fix(1219-dynamodb-resolve-host-exception): fix resolve exception, fix…

* fix(1219-dynamodb-resolve-host-exception): fix resolve exception, fix…

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): new DyanmoDB get/put item …

* fix(1219-dynamodb-resolve-host-exception): add scan-item.sh

* fix(1219-dynamodb-resolve-host-exception): add scan-item.sh

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* fix(1219-dynamodb-resolve-host-exception): streamline dynamodb DTOs

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* develop(1219-dynamodb-resolve-host-exception): fix secretsmanager RES…

* develop(1219-dynamodb-resolve-host-exception): fix secretsmanager RES…

* develop(1219-dynamodb-resolve-host-exception): fix secretsmanager ent…

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* develop(1219-dynamodb-resolve-host-exception): fix secretsmanager ent…

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* develop(1219-secretsmanager-roattion): fix secretsmanager rotation

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

* develop(1219-secretsmanager-roattion): fix secretsmanager rotation

* develop(1219-secretsmanager-roattion): cleanup secrets entity

* develop(1219-secretsmanager-rotation): add lambda result page

* Merge branch 'main' into 1219-dynamodb-resolve-host-exception

+++ src/dto/include/awsmock/dto/secretsmanager/GetSecretValueResponse.h
@@ -47,7 +47,7 @@ namespace AwsMock::Dto::SecretsManager {
         /**
          * Created date
          */
-        double createdDate = -1;
+        long createdDate = -1;
 
         /**
          * Secret string
@@ -75,7 +75,7 @@ namespace AwsMock::Dto::SecretsManager {
             GetSecretValueResponse r;
             r.name = Core::Json::GetStringValue(v, "Name");
             r.versionId = Core::Json::GetStringValue(v, "VersionId");
-            r.createdDate = Core::Json::GetDoubleValue(v, "CreatedDate");
+            r.createdDate = Core::Json::GetLongValue(v, "CreatedDate");
             r.secretString = Core::Json::GetStringValue(v, "SecretString");
             r.secretBinary = Core::Json::GetStringValue(v, "SecretBinary");
             r.versionId = Core::Json::GetStringValue(v, "VersionId");
